Added customizable display values
